+++ isy-ueberwachung/src/main/java/de/bund/bva/isyfact/ueberwachung/config/NachbarsystemConfigurationProperties.java
@@ -1,0 +1,17 @@
+package de.bund.bva.isyfact.ueberwachung.config;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import de.bund.bva.isyfact.ueberwachung.nachbarsystemcheck.model.Nachbarsystem;
+
+/** Enthält die Nachbarsysteme, die mittels des Nachbarsystemchecks überprüft werden. */
+public class NachbarsystemConfigurationProperties {
+
+    /** Alle konfigurierten Nachbarsysteme. */
+    private final Map<String, Nachbarsystem> nachbarsysteme = new HashMap<>();
+
+    public Map<String, Nachbarsystem> getNachbarsysteme() {
+        return nachbarsysteme;
+    }
+}