+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - `IFS-334`: Hinzufügen eines Profils für Jacoco und Anpassung der Surefire Config
 - `IFS-410`: Checkstyle Plugin auf Version 3.1.0 angehoben
 - `IFS-454`: Maven Enforcer Rule: dependency-convergence ergänzt.
+- `IFS-436`: Entkopplung Health-Endpoint und hinzufügen Nachbarsystem-Check
+